chore: 디렉토리 및 패키지명 변경
