+++ src/main/java/com/hanh/domain/habit/HabitRepository.java
@@ -1,0 +1,14 @@
+package com.hanh.domain.habit;
+
+import com.hanh.domain.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface HabitRepository extends JpaRepository<Habit, Long> {
+
+    List<Habit> findAllByUser(User user);
+
+}
