+++ src/main/java/com/hanh/domain/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.hanh.domain.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByNickname(String nickname);
+}
