+++ src/main/java/com/hanh/exception/ExceptionHandler.java
@@ -1,0 +1,2 @@
+package com.hanh.exception;public class ExceptionHandler {
+}
