--- src/main/java/hanh/demo/DemoApplication.java
@@ -1,19 +1,0 @@
-package hanh.demo;
-
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@EnableBatchProcessing
-@SpringBootApplication
-@EnableScheduling
-@EnableAsync
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
-	}
-
-}
