--- src/main/java/hanh/demo/common/config/SchedulerConfig.java
@@ -1,26 +1,0 @@
-package hanh.demo.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.scheduling.annotation.SchedulingConfigurer;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.config.ScheduledTaskRegistrar;
-import org.springframework.stereotype.Component;
-
-@Configuration
-class SchedulerConfig implements SchedulingConfigurer {
-
-    @Override
-    public void configureTasks(ScheduledTaskRegistrar taskRegistrar) {
-        ThreadPoolTaskScheduler threadPoolTaskScheduler = new ThreadPoolTaskScheduler();
-
-        threadPoolTaskScheduler.setPoolSize(5);
-        threadPoolTaskScheduler.setThreadGroupName("scheduler thread pool");
-        threadPoolTaskScheduler.setThreadNamePrefix("scheduler-thread-");
-        threadPoolTaskScheduler.initialize();
-
-        taskRegistrar.setTaskScheduler(threadPoolTaskScheduler);
-    }
-
-}
