--- src/main/java/hanh/demo/exception/CustomException.java
@@ -1,17 +1,0 @@
-package hanh.demo.exception;
-
-public class CustomException extends RuntimeException{
-
-    private static final long serialVersionUID = 1L;
-
-    private ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-}
