--- src/main/java/hanh/demo/exception/DataNotFoundException.java
@@ -1,9 +1,0 @@
-package hanh.demo.exception;
-
-public class DataNotFoundException extends CustomException{
-    private static final long serialVersionUID = -211667454324594101L;
-
-    public DataNotFoundException() {
-        super(ErrorCode.NOT_FOUND);
-    }
-}
