--- src/main/java/hanh/demo/exception/ExceptionHandler.java
@@ -1,2 +1,0 @@
-package hanh.demo.exception;public class ExceptionHandler {
-}
