--- src/main/java/hanh/demo/user/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package hanh.demo.user.repository;
-
-import hanh.demo.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByNickname(String nickname);
-}
