--- src/test/java/hanh/demo/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package hanh.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
